Move path extension function to :core:data
